requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
aiosqlite>=0.18.0
alembic>=1.11.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.11.0
httpx>=0.24.0
pytest>=8.4.0
pytest-asyncio>=0.20.3
anyio>=4.0.0
slowapi>=0.1.6
python-multipart>=0.0.6
typing-extensions>=4.0.0
